Clean up tropo correction routine
